# openapi.yaml
openapi: 3.0.3git 
info:
  title: DT SMS Gateway API
  version: 1.0.0
  description: |
    API REST pour l'envoi de SMS en masse et individuel via DT SMS Gateway.
    
    ## Authentification
    Toutes les requêtes API (sauf health check) nécessitent une clé API valide dans l'en-tête `Authorization: Bearer YOUR_API_KEY`.
    
    ## Limites
    - **Quotas journaliers/mensuels** : Définis par client
    - **Rate limiting** : Selon la configuration client
    - **Taille des messages** : 160 caractères par SMS standard
    
    ## Formats de numéros
    - **Format Djibouti** : `77XXXXXX` ou `+25377XXXXXX`
    - **International** : `+[code_pays][numéro]`
    
    ## Statuts SMS
    - `pending` : En attente d'envoi
    - `sent` : Envoyé au gateway
    - `delivered` : Livré au destinataire  
    - `failed` : Échec d'envoi
  contact:
    name: KASSIM MOUSSA ALI
    url: https://mysms.djiboutitelecom.dj/support
    email: kassim.moussa@djibtel.dj
  license:
    name: Proprietary
    url: https://mysms.djiboutitelecom.dj/license
servers:
  - url: https://mysms.djiboutitelecom.dj/api
    description: Production 
  - url: http://localhost:8000/api
    description: Développement local
security:
  - ApiKeyAuth: []
paths:
...
  /v1/sms/send:
    post:
      tags:
        - SMS Individual
      summary: Envoyer un SMS individuel
      description: |
        Envoie un SMS à un destinataire unique.
        
        **Limite** : Soumis aux quotas journaliers/mensuels du client.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - to
                - message
              properties:
                to:
                  type: string
                  description: "Numéro du destinataire (format Djibouti: 77XXXXXX ou +25377XXXXXX)"
                  example: "77166677"
                  pattern: "^(\\+253)?77[0-9]{6}$"
                message:
                  type: string
                  description: "Contenu du message SMS"
                  example: "Bonjour, ceci est un message test."
                  maxLength: 160
                from:
                  type: string
                  description: "ID expéditeur (texte jusqu'à 11 caractères ou numéro)"
                  example: "DT SMS"
                  maxLength: 11
                async:
                  type: boolean
                  description: "Envoyer de façon asynchrone via queue"
                  default: false
                  example: false
            examples:
              sms_simple:
                summary: SMS simple
                value:
                  to: "77166677"
                  message: "Votre code de vérification est: 123456"
              sms_avec_expediteur:
                summary: SMS avec expéditeur personnalisé
                value:
                  to: "+25377166677"
                  message: "Bienvenue chez DT SMS Gateway!"
                  from: "DT SMS"
      responses:
        '200':
          description: SMS envoyé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "SMS sent successfully"
                  data:
                    type: object
                    properties:
                      sms_id:
                        type: integer
                        example: 12345
                        description: "ID unique du SMS"
                      to:
                        type: string
                        example: "+25377166677"
                      from:
                        type: string
                        example: "DT SMS"
                      status:
                        type: string
                        enum: [pending, sent, delivered, failed]
                        example: "sent"
                      created_at:
                        type: string
                        format: date-time
                        example: "2025-09-24T10:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '422':
          $ref: '#/components/responses/ValidationError'
  /v1/sms:
    get:
      tags:
        - SMS Individual
      summary: Lister les SMS envoyés
      description: Récupère l'historique des SMS avec filtrage et pagination
      parameters:
        - name: status
          in: query
          description: Filtrer par statut
          schema:
            type: string
            enum: [pending, sent, delivered, failed]
        - name: from_date
          in: query
          description: Date de début (format ISO 8601)
          schema:
            type: string
            format: date-time
            example: "2025-09-01T00:00:00Z"
        - name: to_date
          in: query
          description: Date de fin (format ISO 8601)
          schema:
            type: string
            format: date-time
            example: "2025-09-30T23:59:59Z"
        - name: to
          in: query
          description: Filtrer par numéro destinataire
          schema:
            type: string
            example: "77166677"
        - name: page
          in: query
          description: Numéro de page
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          description: Éléments par page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Liste des SMS
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SmsMessage'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /v1/sms/{id}/status:
    get:
      tags:
        - SMS Individual  
      summary: Statut d'un SMS
      description: Récupère le statut détaillé d'un SMS spécifique
      parameters:
        - name: id
          in: path
          required: true
          description: ID du SMS
          schema:
            type: integer
            example: 12345
      responses:
        '200':
          description: Statut du SMS
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/SmsMessage'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Clé API à inclure dans l'en-tête Authorization.
        Format: `Authorization: Bearer YOUR_API_KEY`
  schemas:
    SmsMessage:
      type: object
      properties:
        id:
          type: integer
          example: 12345
        to:
          type: string
          example: "+25377166677"
        from:
          type: string
          example: "DT SMS"
        content:
          type: string
          example: "Votre message SMS"
        status:
          type: string
          enum: [pending, sent, delivered, failed]
          example: "delivered"
        error_message:
          type: string
          example: null
          nullable: true
        sent_at:
          type: string
          format: date-time
          example: "2025-09-24T10:30:00Z"
          nullable: true
        delivered_at:
          type: string
          format: date-time
          example: "2025-09-24T10:30:05Z"
          nullable: true
        created_at:
          type: string
          format: date-time
          example: "2025-09-24T10:30:00Z"
...
    Pagination:
      type: object
      properties:
        current_page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        last_page:
          type: integer
          example: 8
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Une erreur est survenue"
        errors:
          type: object
          additionalProperties: 
            type: array
            items:
              type: string
          example:
            to: ["Le champ to est requis."]
            message: ["Le message ne peut pas dépasser 160 caractères."]
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  message:
                    example: "Requête invalide"
    Unauthorized:
      description: Non autorisé - Clé API invalide ou manquante
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  message:
                    example: "Clé API invalide ou manquante"
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  message:
                    example: "Ressource non trouvée"
    ValidationError:
      description: Erreur de validation
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  message:
                    example: "Validation failed"
    RateLimited:
      description: Limite de débit dépassée
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  message:
                    example: "Daily SMS quota exceeded"
          examples:
            daily_quota_exceeded:
              summary: Quota journalier dépassé
              value:
                success: false
                message: "Daily SMS quota exceeded"
                limit: 1000
                used_today: 1000
                remaining: 0
            monthly_quota_exceeded:
              summary: Quota mensuel dépassé  
              value:
                success: false
                message: "Monthly SMS quota exceeded"
                limit: 25000
                used_this_month: 25000
                remaining: 0
...